[sonarqube] allow overriding app fullname
Allows to override the `fullname` used in various places. Today, a
helper template function determines the format of the name, and it can't
be overridden. This is inconvenient and unwanted in many cases (e.g.
when tooling maps the app name to a hostname—I want
`sonarqube.example.org`, but today I can only get
`sonarqube-sonarqube.example.org`). It's generally better to allow the
fullname to be explicitly overridden if so desired.