Fix plugin download issue for StatefulSet chart (#113)
* Fix plugin download issue for StatefulSet chart

Downloading of plugins fail since the plugin init container is not allowed to write the downloaded plugin into the download directory. If setting values for plugins: runAsUser and runAsGroup to 0 the download works. But, when SonarQube container starts up it doesn't have the permissions to move the plugins in to the correct folder. 

This fix solve the issue so the default user-id 1000 can download the plugin and SonarQube container then have the permissions to move the downloaded files. This issue has been fixed in the deployment.yaml file but must have been missed for the stateful set.