Add helper to flatten hashmap for sonarProperties
This adds a helper to flatten sonarProperties dictionary so that the
key is formatted properly. Without this helper the values for nested
keys get improperly strigified as things like this:

```
sonar=map[governance:map[report:map[project:map[branch:map[frequency:Daily]]
```

This joins the strings with dots as expected and sets the values
correctly.