Use exec command from liveness probe for startup probe by default
The purpose of the startup probe is to tell kubernetes when an application is started. As soon as the probe is successful kubernetes stops checking it and starts checking the liveness and readiness probe.
With the current configuration it sometimes happens that the startup probe is successful but the subsequent liveness probes are not yet and kubernetes kills the pod. By aligning the startup probe with the liveness probe, it should avoid these situations and keep the sonar pods running smoothly.