feat: allow extraVolumeMounts and additional curl options for plugin-install
This Commit introduces two new helm values for configuring
extraVolumeMounts and additonal curl options for the install-plugins
initContainer. The values can be used to mount additonal CABundles into
the install-plugins container. With curlOpts the mounted CABundle can be
used in the shell script that invokes curl to download the plugins.

For example:

plugins.curlOpts: "--cacert=/certs/myCABundle.crt"
plugins.extraVolumeMounts:
  - mountPath: /certs/myCABundle.crt
    subPath: myCABundle.crt
    name: my-ca-bundle